+++ rlclientlib/multistep_loop.cc
@@ -16,7 +16,7 @@ int multistep_loop::request_episodic_decision(str_view event_id, str_view previo
 {
   INIT_CHECK();
   return _pimpl->request_episodic_decision(event_id.str, previous_id.str,
-      {context_json.str, static_cast<size_t>(context_json.size)}, flags, resp, episode, status);
+      {context_json.str, context_json.size}, flags, resp, episode, status);
 }
 
 int multistep_loop::report_outcome(str_view episode_id, str_view event_id, float outcome, api_status* status)
